test program 1: file with all possible opcodes
This file is artificially generated (i.e. not a meaningful COP4 program) to include all possible 1 and 2 byte opcodes of the COP4 processor. 

The .bin is the input to the disassembler, while .txt is the output. 

To test the disassembler you disassemble the binary file (.bin) and then compare the output with the text file (.txt)