Formatting changes and .WORD directive
Some commands have an implicit argument (which is equal to 0). Do we have X instead of X 0.
I have changed the disassembler to follow this convention.

Got rid of UNDEF for unknown or illegal opcodes, replacing them with the .WORD   directive which reserves space.
The two say the same thing, but the new version is understandable by the assembler.